Basic STOMP Server (SI)
Core STOMP infrastructure (converter, handler) is in o.s.stomp and has no SI dependencies.

Currently, it only supports CONNECT, SUBSCRIBE, UNSUBSCRIBE, DISCONNECT.

There's an abstract SI StompSubscriptionMessageProducer and a test
TestStompSubscriptionMessageProducer that emits 'Hello, world!' every
n milliseconds to open subscriptions.

This is configured into a test case StompOverWebSocketTests which
configures the message producer to send to its subscriptions every
2 seconds.

Run StompOverWebSocketTests and open stomp.html in a websocket-capable
browser (tested with Chrome on linux).

Currently the SI message payload is a StompMessage (headers and payload)
but there *is* a pair of STOMP transformers.